Initial rewrite pymatcher -- now it's a ctrlp matcher that does sane ranking.
Not much of the original code is left; I mostly used it as a shim.

Documentation hasn't yet been updated.  And most of the ctrl-p features
aren't supported.

Discussion is with original author as to whether the license should have
a copyright assignment to Google/Angular.js -- that seems like it might
be a copy/paste error.